Refactored Haka/Shibboleth login attributes
Use EPPN for user identification, mail is an optional parameter not used yet,
i.e. it's not atm. guaranteed that mail delivery will work if user login with
Shibboleth. Then again, nothing guarantees that the mail attribute has a valid
value even if it exists.

Opened issue #264 for this redesign