multiple network support and use modern openstack clients
Add the ability to add net UUID to support running pebbles in a tenant
with multiple networks. Specifying the network for OpenStackClient is
not yet implemented, though.

Refactor use of python-novaclient to python-novaclient,
python-cinderclient and python-neutronclient. We still use v2 API
though.

Clean up some docs etc. in the process.