instance running logs for release-5
Add support for triggering a fetch for instance running logs by the
worker. The request for fetching is done through a new boolean field
'log_fetch_pending' in Instance. After uploading the logs the worker
will set it back to false.

changes:
- add log_fetch_pending to Instance
- add support for patching log_fetch_pending in InstanceView
- change Worker to take running instances with log_fetch_pending into
  account when processing updates
- change base_driver to coordinate log fetching
  - base_driver calls subclass to provide the logs
  - base_driver uploads the logs and updates Instance.log_fetch_pending
- implement pod log fetching in kubernetes_driver
- change PBClient.patch to also work with both form and json data

Bonus: remove logger.py, that is not used anymore for log uploading