workspace expiry timestamp
Add expiry_ts to Workspace and rename created_at to create_ts. These
fields are implemented as hybrid fields, so that the database
representation is datetime (timestamp without timezone in postgresql)
for human readable data and the API hands out the timestamps as
integers, to be converted to human readable form according to browser
locale.

Also: move workspace_fields from commons.py to the only referencing
module workspaces.py. Update dateutil in requirements.txt. Add a check
for expiry_ts in the unit tests.