move cluster binding to workspace
Move the cluster selection for instances to workspace from environment
templates. Workspace is naturally bound to a cluster, because workspace
data is located in the cluster.

Add config option for setting the default cluster for new workspaces.

Environment.full_config is removed as redundant, and the method of
getting full_config is renamed to the current purpose of rendering
provisioning config for instances.

Redundant assignment to instance.can_update_connectivity is removed.