add image pull policy to deployments
Sometimes customers reuse the same image tag for updated images as
well. In this case, we need to be able to force checking if the image
on the node is actually up-to-date. Cache will still be honored, so even
if we pull we don't transfer existing layers again.

Why not make 'Always' policy the default then? In case of registry being
unavailable for any reason, we would not be able to launch any sessions.
Docker Hub also counts checking the image metadata against its rate
limits, and exceeding the limit will cause a session startup failure.